requirements.txt
streamlit==1.29.0
psycopg2-binary==2.9.9
python-dotenv==1.0.0
SQLAlchemy==2.0.23 